requirements.txt
asgiref==3.5.2
Django==3.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
mysqlclient==2.1.1
PyJWT==2.6.0
pytz==2022.7
sqlparse==0.4.3
